changed popsize param from int to std::size_t in the BitSet interface

+++ src/graph/tests/test_clq_func.cpp
@@ -82,7 +82,7 @@ TEST_F(CliqueCommonTest, find_clique) {
 
 TEST_F(CliqueCommonTest, find_clique_2) {
 
-	typename ugraph::_bbt bb{ (int)ug.size(), {2, 3, 4, 5}};
+	typename ugraph::_bbt bb{ ug.size(), {2, 3, 4, 5}};
 	std::vector<int> clq;
 
 
@@ -120,7 +120,7 @@ TEST_F(CliqueCommonTest, find_clique_2) {
 
 TEST_F(CliqueCommonTest, is_iset) {
 
-	typename ugraph::_bbt bb{ static_cast<int>(ug.size()), { 1, 3, 4} };
+	typename ugraph::_bbt bb{ ug.size(), { 1, 3, 4} };
 
 	///////////////////
 	//bitset of vertices
@@ -164,7 +164,7 @@ TEST_F(CliqueCommonTest, is_iset) {
 
 TEST_F(CliqueCommonTest, is_clique) {
 
-	typename ugraph::_bbt bb{ static_cast<int>(ug.size()), { 0, 4 } };
+	typename ugraph::_bbt bb{ ug.size(), { 0, 4 } };
 
 	///////////////////
 	//bitset of vertices
@@ -226,7 +226,7 @@ TEST(Clique, find_clique_max_deg) {
 	ug.add_edge(3, 4);
 
 	std::vector<int> clq;
-	typename ugraph::_bbt bbsg{ static_cast<int>(ug.size()), { 0, 1, 2, 3, 4, 5} };
+	typename ugraph::_bbt bbsg{ ug.size(), { 0, 1, 2, 3, 4, 5} };
 
 	/////////////////////////////////////////////////////////
 	int nV = gfunc::clq::find_clique_max_deg<decltype(ug), false>(ug, clq, bbsg);
@@ -251,7 +251,7 @@ TEST(Clique, find_clique_in_pool) {
 	ug.add_edge(3, 4);
 
 	std::vector<int> clq;
-	typename ugraph::_bbt bbsg{ static_cast<int>(ug.size()), { 0, 1, 2, 3, 4, 5} };
+	typename ugraph::_bbt bbsg{ ug.size(), { 0, 1, 2, 3, 4, 5} };
 
 	/////////////////////////////////////////////////////////
 	int nV = gfunc::clq::find_clique_from_pool(ug, clq, bbsg);
@@ -275,7 +275,7 @@ TEST(Clique, find_max_clique_lb) {
 	ug.add_edge(3, 4);
 
 	std::vector<int> clq;
-	typename ugraph::_bbt bbsg{ static_cast<int>(ug.size()), { 0, 1, 2, 3, 4, 5} };
+	typename ugraph::_bbt bbsg{ ug.size(), { 0, 1, 2, 3, 4, 5} };
 
 	/////////////////////////////////////////////////////////
 	int nV = gfunc::clq::find_clique_lb(ug);
@@ -300,7 +300,7 @@ TEST(Clique, SEQ_coloring) {
 
 
 	std::vector<int> clq;
-	typename ugraph::_bbt bbsg{ static_cast<int>(ug.size()), { 0, 1, 2, 3, 4, 5} };
+	typename ugraph::_bbt bbsg{ ug.size(), { 0, 1, 2, 3, 4, 5} };
 
 
 	////////////////////////////////////////
